added delete mutation

+++ app/GraphQL/Mutations/DeleteMediaMutation.php
@@ -11,15 +11,28 @@ class DeleteMediaMutation
     public function __construct(private MediaService $mediaService) {}
 
     /**
-     * @param array{id: string} $args
+     * @param array{ids: array<string>} $args
+     * @return array{success: bool, deletedCount: int}
      */
-    public function __invoke(null $_, array $args): bool
+    public function __invoke(null $_, array $args): array
     {
         Validator::make($args, [
-            'id' => ['required', 'exists:media,id'],
+            'ids' => ['required', 'array', 'min:1'],
+            'ids.*' => ['required', 'string', 'exists:media,id'],
         ])->validate();
 
-        $media = Media::findOrFail($args['id']);
-        return $this->mediaService->deleteMedia($media);
+        $deletedCount = 0;
+        foreach ($args['ids'] as $id) {
+            $media = Media::find($id);
+            if ($media) {
+                $this->mediaService->deleteMedia($media);
+                $deletedCount++;
+            }
+        }
+
+        return [
+            'success' => $deletedCount > 0,
+            'deletedCount' => $deletedCount,
+        ];
     }
 }
